Move libminifloat and liblogintpack to DataFormats/Math
