+++ DataFormats/Math/test/testRunner.cpp
@@ -1,0 +1,1 @@
+#include "Utilities/Testing/interface/CppUnit_testdriver.icpp"
