+++ DataFormats/PatCandidates/src/CovarianceParameterization.cc
@@ -1,6 +1,6 @@
 #include "DataFormats/PatCandidates/interface/CovarianceParameterization.h"
-#include "DataFormats/PatCandidates/interface/liblogintpack.h"
-#include "DataFormats/PatCandidates/interface/libminifloat.h"
+#include "DataFormats/Math/interface/liblogintpack.h"
+#include "DataFormats/Math/interface/libminifloat.h"
 #include "FWCore/ParameterSet/interface/FileInPath.h"
 #include <boost/format.hpp>
 #include <iostream>
