+++ PhysicsTools/PatAlgos/plugins/PATVertexSlimmer.cc
@@ -11,7 +11,7 @@
 #include "FWCore/Framework/interface/MakerMacros.h"
 #include "DataFormats/VertexReco/interface/Vertex.h"
 #include "DataFormats/Common/interface/ValueMap.h"
-#include "DataFormats/PatCandidates/interface/libminifloat.h"
+#include "DataFormats/Math/interface/libminifloat.h"
 
 namespace pat {
   class PATVertexSlimmer : public edm::global::EDProducer<> {
